基本完成ArticleService

+++ src/main/java/top/wetor/service/impl/ArticleServiceImpl.java
@@ -1,6 +1,5 @@
 package top.wetor.service.impl;
 
-import org.apache.ibatis.jdbc.Null;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import top.wetor.dao.ArticleDao;
@@ -8,8 +7,9 @@ import top.wetor.dao.TagDao;
 import top.wetor.dao.TagMapDao;
 import top.wetor.domain.Article;
 import top.wetor.service.IArticleService;
-import top.wetor.utils.ArticleMode;
+import top.wetor.utils.ArticleContentShow;
 import top.wetor.domain.*;
+import top.wetor.utils.ArticleStateShow;
 
 import java.util.List;
 @Service("articleService")
@@ -21,6 +21,37 @@ public class ArticleServiceImpl implements IArticleService {
     @Autowired
     private TagMapDao tagMapDao;
 
+    private Integer getContentShow(ArticleContentShow content){
+        switch (content){
+            case MINI: return 0;
+            case SIMPLE: return 1;
+            case COMPLETE: return 2;
+        }
+        return 0;
+    }
+    private Integer getStateShow(ArticleStateShow state){
+        switch (state){
+            case NORMAL: return 0;
+            case TOP: return 1;
+            case HIDDEN: return 2;
+            case DELETE: return 3;
+            case USER: return 4;
+            case ADMIN: return 5;
+        }
+        return 0;
+    }
+    private ArticleStateShow getStateShowByArticle(Article article){
+        if(article.getDelete()) return ArticleStateShow.DELETE;
+        if(article.getHidden()) return ArticleStateShow.HIDDEN;
+        if(article.getTop()) return ArticleStateShow.TOP;
+        return ArticleStateShow.NORMAL;
+    }
+    private Integer getBeanType(Object bean){
+        if(bean instanceof User) return 0;
+        if(bean instanceof Group) return 1;
+        if(bean instanceof Tag) return 2;
+        return null;
+    }
     /**
      * 把文章加入数据库,并更新标签和标签映射表
      * @param article 文章
@@ -38,9 +69,15 @@ public class ArticleServiceImpl implements IArticleService {
      */
     @Override
     public void deleteArticleById(Integer id) {
-        Article article = articleDao.selectArticleById(id,0);//极简查询
+
+        Article article = articleDao.selectArticleById(id,getContentShow(ArticleContentShow.MINI));//极简查询
+
         Article updateArticle=new Article();
+        updateArticle.setId(article.getId());
         updateArticle.setDelete(true);
+        updateArticle.setTop(false);
+        updateArticle.setHidden(false);
+
         articleDao.updateArticle(updateArticle);
         tagDao.deleteTagByString(article.getTags());
         tagMapDao.deleteTagMapByArticleId(article.getId());
@@ -51,7 +88,16 @@ public class ArticleServiceImpl implements IArticleService {
      */
     @Override
     public void deleteArticleRestoreById(Integer id) {
+        Article article = articleDao.selectArticleById(id,getContentShow(ArticleContentShow.MINI));//极简查询
+        if(getStateShowByArticle(article)!=ArticleStateShow.DELETE){
+            //TODO 抛出异常
+            return;
+        }
+        article.setDelete(false);
 
+        articleDao.updateArticle(article);
+        tagDao.updateTagByString(article.getTags());
+        tagMapDao.insertTagMapByArticle(article.getId(),article.getTags());
     }
     /**
      * 通过id完全删除文章
@@ -59,81 +105,54 @@ public class ArticleServiceImpl implements IArticleService {
      */
     @Override
     public void deleteArticleCompleteById(Integer id) {
-        Article article = articleDao.selectArticleById(id,0);//极简查询
+        Article article = articleDao.selectArticleById(id,getContentShow(ArticleContentShow.MINI));//极简查询
+        if(getStateShowByArticle(article)!=ArticleStateShow.DELETE){
+            //TODO 抛出异常
+            return;
+        }
         articleDao.deleteArticleById(id);
-        tagDao.deleteTagByString(article.getTags());
-        tagMapDao.deleteTagMapByArticleId(article.getId());
     }
 
     @Override
     public void updateArticle(Article article) {
+        Article oldArticle = articleDao.selectArticleById(article.getId(),getContentShow(ArticleContentShow.MINI));//极简查询
+        tagDao.deleteTagByString(oldArticle.getTags());
+        tagMapDao.deleteTagMapByArticleId(oldArticle.getId());
+
+
         articleDao.updateArticle(article);
+        tagDao.updateTagByString(article.getTags());
+        tagMapDao.insertTagMapByArticle(article.getId(),article.getTags());
     }
 
     @Override
-    public Article selectArticleById(Integer id, ArticleMode mode) {
-        switch (mode){
-            case MINI:
-                return articleDao.selectArticleById(id,0);
-            case SIMPLE:
-                return articleDao.selectArticleById(id,1);
-            case COMPLETE:
-                return articleDao.selectArticleById(id,2);
-        }
-        return null;
-
+    public Article selectArticleById(Integer id, ArticleContentShow content) {
+        return articleDao.selectArticleById(id,getContentShow(content));
     }
 
     @Override
-    public Integer selectArticleCount() {
-        return articleDao.selectArticleCount();
+    public Integer selectArticleCount(ArticleStateShow state) {
+        return articleDao.selectArticleCount(getStateShow(state));
     }
 
     @Override
-    public List<Article> selectArticlePage(Integer begin, Integer number, ArticleMode mode, Boolean asc) {
-        switch (mode){
-            case MINI:
-                return articleDao.selectArticlePage(begin, number, 0, asc);
-            case SIMPLE:
-                return articleDao.selectArticlePage(begin, number, 1, asc);
-            case COMPLETE:
-                return articleDao.selectArticlePage(begin, number, 2, asc);
-        }
-        return null;
+    public List<Article> selectArticlePage(Integer begin, Integer number, ArticleContentShow content, Boolean asc, ArticleStateShow state) {
+        return articleDao.selectArticlePage(begin, number, getContentShow(content), asc, getStateShow(state));
     }
 
     @Override
-    public List<Article> selectArticleAll(ArticleMode mode, Boolean asc) {
-        switch (mode){
-            case MINI:
-                return articleDao.selectArticleAll(0, asc);
-            case SIMPLE:
-                return articleDao.selectArticleAll(1, asc);
-            case COMPLETE:
-                return articleDao.selectArticleAll(2, asc);
-        }
-        return null;
+    public List<Article> selectArticleAll(ArticleContentShow content, Boolean asc, ArticleStateShow state) {
+        return articleDao.selectArticleAll(getContentShow(content), asc, getStateShow(state));
     }
 
     @Override
-    public List<Article> selectArticleAllByBean(Object bean, ArticleMode mode, Boolean asc) {
-        if(bean instanceof User)
-            return articleDao.selectArticleAllByType(0, ((User)bean).getId(), 0, asc);
-        if(bean instanceof Group)
-            return articleDao.selectArticleAllByType(1, ((Group)bean).getId(), 1, asc);
-        if(bean instanceof Tag)
-            return articleDao.selectArticleAllByType(2, ((Tag)bean).getId(), 2, asc);
-        return null;
+    public List<Article> selectArticleAllByBean(Object bean, ArticleContentShow content, Boolean asc, ArticleStateShow state) {
+        return articleDao.selectArticleAllByType(getBeanType(bean), ((BaseBean)bean).getId(), getContentShow(content), asc, getStateShow(state));
     }
 
     @Override
-    public List<Article> selectArticlePageByBean(Object bean, Integer begin, Integer number, ArticleMode mode, Boolean asc) {
-        if(bean instanceof User)
-            return articleDao.selectArticlePageByType(0, ((User)bean).getId(), begin, number, 0, asc);
-        if(bean instanceof Group)
-            return articleDao.selectArticlePageByType(1, ((Group)bean).getId(), begin, number,1, asc);
-        if(bean instanceof Tag)
-            return articleDao.selectArticlePageByType(2, ((Tag)bean).getId(), begin, number, 2, asc);
-        return null;
+    public List<Article> selectArticlePageByBean(Object bean, Integer begin, Integer number, ArticleContentShow content, Boolean asc, ArticleStateShow state) {
+        return articleDao.selectArticlePageByType(getBeanType(bean), ((BaseBean)bean).getId(), begin, number, getContentShow(content), asc, getStateShow(state));
+
     }
 }
